refactor: Admin 관련 파일들 디렉토리 위치 수정
